@@ -1,16 +1,4 @@
-import Vue from 'vue';
-import App from './App.vue';
-import router from './router.js';
-import store from './store/index';
-import VueAwesomeSwiper from 'vue-awesome-swiper';
-import 'swiper/swiper-bundle.css'
-Vue.use(VueAwesomeSwiper);
-Vue.use({
-  router,
-  store
-})
-Vue.config.productionTip = false
+import { createApp } from 'vue'
+import App from './App.vue'
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
